Moved TSVDatasetManager to `org.cirdles.topsoil.app.utils`
